Add IAccountManager and ITrasactionManager, add transaction validation with ITransactionValidator
