+++ Project/FinanceManager/FinanceManager.Core/Services/Abstractions/ITransactionValidator.cs
@@ -1,0 +1,7 @@
+﻿using FinanceManager.Core.DataTransferObjects.Commands;
+
+namespace FinanceManager.Core.Services.Abstractions;
+public interface ITransactionValidator
+{
+    public void Validate(PutTransactionDto command);
+}
